requirements.txt
absl-py==2.0.0
anyio==4.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
fqdn==1.5.1
gmpy2==2.1.2
google-auth==2.24.0
google-auth-oauthlib==1.1.0
grpcio==1.59.3
idna==3.6
importlib-metadata==7.0.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.18.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_server==2.11.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterlab-vim==4.1.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.8.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook_shim==0.2.3
numpy==1.26.2
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
rsa==4.9
Send2Trash==1.8.2
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.2
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.3.3
traitlets==5.14.0
triton==2.1.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
typing-utils==0.1.0
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.12
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.42.0
zipp==3.17.0
